# Reduce startup overhead during default key binding handling