--- a/osu.Game/Input/RealmKeyBindingStore.cs
+++ b/osu.Game/Input/RealmKeyBindingStore.cs
@@ -46,49 +46,50 @@ namespace osu.Game.Input
         }
 
         /// <summary>
-        /// Register a new type of <see cref="KeyBindingContainer{T}"/>, adding default bindings from <see cref="KeyBindingContainer.DefaultKeyBindings"/>.
+        /// Register all defaults for this store.
         /// </summary>
         /// <param name="container">The container to populate defaults from.</param>
-        public void Register(KeyBindingContainer container) => insertDefaults(container.DefaultKeyBindings);
-
-        /// <summary>
-        /// Register a ruleset, adding default bindings for each of its variants.
-        /// </summary>
-        /// <param name="ruleset">The ruleset to populate defaults from.</param>
-        public void Register(RulesetInfo ruleset)
-        {
-            var instance = ruleset.CreateInstance();
-
-            foreach (var variant in instance.AvailableVariants)
-                insertDefaults(instance.GetDefaultKeyBindings(variant), ruleset.ID, variant);
-        }
-
-        private void insertDefaults(IEnumerable<IKeyBinding> defaults, int? rulesetId = null, int? variant = null)
+        /// <param name="rulesets">The rulesets to populate defaults from.</param>
+        public void Register(KeyBindingContainer container, IEnumerable<RulesetInfo> rulesets)
         {
             using (var usage = realmFactory.GetForWrite())
             {
-                // compare counts in database vs defaults
-                foreach (var defaultsForAction in defaults.GroupBy(k => k.Action))
+                // intentionally flattened to a list rather than querying against the IQueryable, as nullable fields being queried against aren't indexed.
+                // this is much faster as a result.
+                var existingBindings = usage.Realm.All<RealmKeyBinding>().ToList();
+
+                insertDefaults(usage, existingBindings, container.DefaultKeyBindings);
+
+                foreach (var ruleset in rulesets)
                 {
-                    int existingCount = usage.Realm.All<RealmKeyBinding>().Count(k => k.RulesetID == rulesetId && k.Variant == variant && k.ActionInt == (int)defaultsForAction.Key);
-
-                    if (defaultsForAction.Count() <= existingCount)
-                        continue;
-
-                    foreach (var k in defaultsForAction.Skip(existingCount))
-                    {
-                        // insert any defaults which are missing.
-                        usage.Realm.Add(new RealmKeyBinding
-                        {
-                            KeyCombinationString = k.KeyCombination.ToString(),
-                            ActionInt = (int)k.Action,
-                            RulesetID = rulesetId,
-                            Variant = variant
-                        });
-                    }
+                    var instance = ruleset.CreateInstance();
+                    foreach (var variant in instance.AvailableVariants)
+                        insertDefaults(usage, existingBindings, instance.GetDefaultKeyBindings(variant), ruleset.ID, variant);
                 }
 
                 usage.Commit();
+            }
+        }
+
+        private void insertDefaults(RealmContextFactory.RealmUsage usage, List<RealmKeyBinding> existingBindings, IEnumerable<IKeyBinding> defaults, int? rulesetId = null, int? variant = null)
+        {
+            // compare counts in database vs defaults for each action type.
+            foreach (var defaultsForAction in defaults.GroupBy(k => k.Action))
+            {
+                // avoid performing redundant queries when the database is empty and needs to be re-filled.
+                int existingCount = existingBindings.Count(k => k.RulesetID == rulesetId && k.Variant == variant && k.ActionInt == (int)defaultsForAction.Key);
+
+                if (defaultsForAction.Count() <= existingCount)
+                    continue;
+
+                // insert any defaults which are missing.
+                usage.Realm.Add(defaultsForAction.Skip(existingCount).Select(k => new RealmKeyBinding
+                {
+                    KeyCombinationString = k.KeyCombination.ToString(),
+                    ActionInt = (int)k.Action,
+                    RulesetID = rulesetId,
+                    Variant = variant
+                }));
             }
         }