The main overhead here was from the `.Count` calls, which seem to be very unoptimised. Kind of makes sense since we aren't indexing any of the columns in question, but still a point of concern. May follow up with realm if we run into this kind of thing again, but in this specific case a local optimisation is warranted regardless, as we are iterating this same collection many times and are *unable* to index nullable fields anyways.

Other surrounding changes are to avoid unnecessary multiple transactions for what is a single operation in all cases.

Testing with `Visual.Online` namespace, as this is most noticeable when run many times (as headless tests do):

Before: 50.8s (54.2s with profiling below)

![20210907 152235 (JetBrains Rider)](https://user-images.githubusercontent.com/191335/132294873-583839fd-7b86-4ef5-b5ec-82b8c1ae1870.png)

After: 35.9s (40.6s with profiling below)

![20210907 152411 (JetBrains Rider)](https://user-images.githubusercontent.com/191335/132294718-6e728555-28e0-4da0-b96d-05598dfdb606.png)